+++ health/micro-ui/web/micro-ui-internals/packages/modules/campaign-manager/src/Module.js
@@ -70,13 +70,14 @@ export const CONSOLE_MDMS_MODULENAME = "HCM-ADMIN-CONSOLE";
  */
 const CampaignModule = ({ stateCode, userType, tenants }) => {
   const tenantId = Digit.ULBService.getCurrentTenantId();
+  const moduleName = Digit.Utils.campaign.getModuleName();
   const { data: BOUNDARY_HIERARCHY_TYPE, isLoading: hierarchyLoading } = Digit.Hooks.useCustomMDMS(
     tenantId,
     CONSOLE_MDMS_MODULENAME,
     [
       {
         name: "HierarchySchema",
-        filter: `[?(@.type=='${window.Digit.Utils.campaign.getModuleName()}')]`,
+        filter: `[?(@.type=='${moduleName}')]`,
       },
     ],
     {
@@ -84,7 +85,7 @@ const CampaignModule = ({ stateCode, userType, tenants }) => {
         return data?.[CONSOLE_MDMS_MODULENAME]?.HierarchySchema?.[0]?.hierarchy;
       },
     },
-    { schemaCode: `${CONSOLE_MDMS_MODULENAME}.HierarchySchema` }
+    { schemaCode: "HierarchySchema" }
   );
 
   const hierarchyData = Digit.Hooks.campaign.useBoundaryRelationshipSearch({ BOUNDARY_HIERARCHY_TYPE, tenantId });