+++ health/micro-ui/web/micro-ui-internals/packages/css/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.3.1 - 2025-2-11
+#### Admin console , Microplan & Payments
+  1. Campaign Manager Module CSS
+  2. Boundary Manager Module CSS
+  3. Microplan Module CSS
+  4. Payments Module CSS
+  5. Other Core Override CSS
+
 ## 0.3.0 - 2024-12-03
 #### Base Admin console & microplan web
   1. Campaign Manager Module CSS